+++ packages/sidecar/CHANGELOG.md
@@ -1,5 +1,12 @@
 # kalguard-sidecar
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - kalguard-core@1.1.1
+
 ## 1.0.1
 
 ### Patch Changes